WorkSpace/sell/src/main/java/com/lt/service/impl/CategoryServiceImpl.java
package com.lt.service.impl;

import java.util.List;


import com.lt.dataobject.ProductCategory;
import com.lt.repository.ProductCategoryRepository;
import com.lt.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * @author lt
 * @date 2018/5/10
 */
@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryRepository repository;


    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return save(productCategory);
    }
}
